fix: Rename series_types/apple/ to series_types/sdk/
