+++ backend/app/constants/series_types/sdk/category_types.py
