+++ backend/app/constants/series_types/sdk/workout_statistics.py
