+++ backend/scripts/generate_healthkit_payload.py
@@ -28,7 +28,7 @@ from pathlib import Path
 from typing import Any
 from uuid import uuid4
 
-from app.constants.series_types.apple import AppleCategoryType, SDKMetricType, SleepPhase
+from app.constants.series_types.sdk import AppleCategoryType, SDKMetricType, SleepPhase
 from app.constants.workout_types.apple_sdk import SDKWorkoutType
 
 # Record types with their HealthKit units and realistic value ranges (min, max)
